refactor upload handling
